refactor: apply the Rider inspection findings (#52)
* refactor(ui): use the primary-constructor parameters directly

The five readonly copies of a primary-constructor parameter in
ChartCursorReader, ChartDeltaCursorReader and ChartRealtimeApplier go;
the parameters are captured as the rest of the UI already does.


* refactor: apply the Rider inspection findings

The five toolbar command methods left dead by the lambda inlining, two
unused usings in TrendChartView, TrendChartViewModel's logger field that
had become a local, the ValueTask GetResult on Serilog's flush, the
seeder's Ctrl+C handler that outlived its CancellationTokenSource, and
two pairs of duplicated if branches.


* fix(seeder): dispose the stop source after the handler is gone


* test: apply the Rider inspection findings

Thirteen unused locals, an unused using, a redundant cast, four redundant
null-forgiving operators, explicit lambda parameter types, a default
argument restated, the two Chart test files' namespace matched to their
folder, and the three _unused disposal locals renamed to the rule.


---------

+++ SemiPlot/SemiPlot.Tests.Data/Integration/FollowRestartTests.cs
@@ -68,7 +68,7 @@ public sealed class FollowRestartTests(PostgresContainerFixture postgresContaine
 			cancellationToken);
 
 		Assert.Equal(
-			(long)(_firstRunRows.Count + secondRunRows.Count),
+			_firstRunRows.Count + secondRunRows.Count,
 			await CountRawRowsAsync(cancellationToken));
 	}
